se agrega las card en vuelos.html

+++ ProyectoViajes/ProyectoViajesApp/models.py
@@ -10,7 +10,7 @@ class Vuelo(models.Model):
     destino = models.CharField(max_length=30)
     ida = models.DateTimeField()
     vuelta = models.DateTimeField()
-    precio = models.FloatField()
+    precio = models.IntegerField()
 
 class Hotel(models.Model):
 
@@ -19,7 +19,7 @@ class Hotel(models.Model):
     descripcion = models.CharField(max_length=150)
     desde = models.DateTimeField()
     hasta = models.DateTimeField()
-    precio = models.FloatField()
+    precio = models.IntegerField()
 
 class Excursion(models.Model):
 
@@ -27,4 +27,4 @@ class Excursion(models.Model):
     ubicacion = models.CharField(max_length=60)
     descripcion = models.CharField(max_length=150)
     duracion = models.DateTimeField()
-    precio = models.FloatField()
+    precio = models.IntegerField()
